desarrollo de Clase Message y ajust de utilitario para manejo de archivos

+++ src/utilities/files.py
@@ -1,11 +1,11 @@
+from control_messages import Mensage
 import os
 import shutil
-
 
 class FileUtils:
 
     @classmethod
-    def Save(cls, name: str, path: str ,data: str) -> bool:
+    def Save(cls, name: str, path: str ,data: str) -> dict:
         """
          Save, Guarda en un archivo la informacion entregada        
         
@@ -17,10 +17,10 @@ class FileUtils:
             data (str): Contiene los datos que se crearan en el archivo
 
         Returns:
-            Bool: Retorna True si se logra crear el archivo, en caso contrario retorna False
+            Dict: Retorna diccionario con mensaje del resultado de la operacion, tipo de resultado (successful,warning) y un estado True si la operacion finalizo con exito o False en caso contrario 
             
         Note:
-            -Si el archivo ya existe en la ruta de destino este adicionara el nuevo dato en una linea nueva
+            - Si el archivo ya existe en la ruta de destino este adicionara el nuevo dato en una linea nueva
         """
         try:
             
@@ -30,27 +30,30 @@ class FileUtils:
             with open(os.path.join(path,name), 'a') as file:
                 file.write(str(data) + '\n')
                 
-            return True        
+            return  Mensage.build_message(id_mesage = 0)        
         except Exception as e:
-            return False
+            return Mensage.build_message(0,str(e.args[1]),e.filename,e.filename2)
     
     @classmethod
     def read_file(cls,path_file: str) -> str: 
         """
-         read_file, lee la informacion contenida en el archivo especificado y lo retorna como un string        
+        read_file, lee la informacion contenida en el archivo especificado y lo retorna como un string        
         
         Args:
             path_file (str): Ruta del archivo a leer
 
         Returns:
-            str: Retorna string con el contenido del archivo leeido
+            str: Retorna string con el contenido del archivo leeido, en caso de un error retorna un valor None
         """
-        with open(path_file) as file:
-            return file.read()
+        try:
+            
+            with open(path_file) as file:
+                return file.read()
+        except Exception as e:
+            return None
     
-          
     @classmethod
-    def move_file(cls,origin_path: str,destination_path: str) -> bool:
+    def move_file(cls,origin_path: str,destination_path: str) -> dict:
         
         """
         move_file, mueve los archivos de la ruta origen a una ruta destino, si el archivo ya existe en la ruta de destino este se remplazara
@@ -61,12 +64,15 @@ class FileUtils:
             origin_path (str): Ruta de destino de los archivos
 
         Returns:
-            bool: Retorna True si los archivos de movieron con exito, en caso contrario se retorna False
+            Dict: Retorna diccionario con mensaje del resultado de la operacion, tipo de resultado (successful,warning) y un estado True si la operacion finalizo con exito o False en caso contrario 
         
         Note:
             -Si el archivo ya existe en la ruta de destino este se remplazara
         """
         try:
+            
+            if not os.path.exists(origin_path):
+                return  Mensage.build_message(id_mesage = 1,part1_mesage = origin_path)
             
             if not os.path.exists(destination_path):
                 os.makedirs(destination_path)
@@ -78,42 +84,12 @@ class FileUtils:
                             os.path.join(destination_path, file)
                             )
 
-            return True        
+            return  Mensage.build_message(id_mesage = 0)          
         except Exception as e:
-            return False
-         
+            return Mensage.build_message(0,str(e.args[1]),e.filename,e.filename2)
     
     @classmethod
-    def copy_file(cls,origin_path: str,destination_path: str) -> bool:
-        """
-        copy_file, copia los archivos de la ruta origen a una ruta destino, si el archivo ya existe en la ruta de destino este se remplazara
-        
-        Args:
-            origin_path (str): Ruta de origen de los archivos
-            
-            origin_path (str): Ruta de destino de los archivos
-
-        Returns:
-            bool: Retorna True si los archivos de movieron con exito, en caso contrario se retorna False
-        
-        """
-        try:
-            if not os.path.exists(destination_path):
-                os.makedirs(destination_path)
-            
-            files = os.listdir(origin_path) 
-            
-            for file in files:
-                shutil.copy(os.path.join(origin_path, file),
-                            os.path.join(destination_path, file)
-                            ) 
-                
-            return True        
-        except Exception as e:
-            return False
-        
-    @classmethod
-    def delete_file(cls,path_file: str) -> bool:
+    def delete_file(cls,path_file: str) -> dict:
         """
         delete_file, elimina el archivo especificado
         
@@ -121,17 +97,21 @@ class FileUtils:
             path_file (str): Ruta del archivo a eliminar
 
         Returns:
-            bool: Retorna True si el archivo se elimino exitosamente, en caso contrario se retorna False
+            Dict: Retorna diccionario con mensaje del resultado de la operacion, tipo de resultado (successful,warning) y un estado True si la operacion finalizo con exito o False en caso contrario 
         
         """
         try:
+            if not os.path.exists(path_file):
+                return  Mensage.build_message(id_mesage = 3,part1_mesage = path_file)
+            
             os.remove(path_file)
-            return True        
+            
+            return  Mensage.build_message(id_mesage = 0)        
         except Exception as e:
-            return False
+            return Mensage.build_message(0,str(e.args[1]),e.filename,e.filename2)
     
     @classmethod
-    def clear_folder(cls,path_folder: str) -> bool:
+    def clear_folder(cls,path_folder: str) -> dict:
         """
         clear_folder, elimina los archivos contenidos dentro de la carpeta especificada
         
@@ -139,18 +119,54 @@ class FileUtils:
             path_folder (str): Ruta de la carpeta a limpiar
 
         Returns:
-            bool: Retorna True si la carpeta se limpio exitosamente, en caso contrario se retorna False
+            Dict: Retorna diccionario con mensaje del resultado de la operacion, tipo de resultado (successful,warning) y un estado True si la operacion finalizo con exito o False en caso contrario 
         """
 
         try:
+            
+            if not os.path.exists(path_folder):
+                return  Mensage.build_message(id_mesage = 2,part1_mesage = path_folder)
+            
             files = os.listdir(path_folder)
             for file in files:
                 path_file = os.path.join(path_folder, file)
                 os.remove(path_file)
-            return True        
+                
+            return  Mensage.build_message(id_mesage = 0) 
+              
         except Exception as e:
-            return False
-    
+            return Mensage.build_message(0,str(e.args[1]),e.filename,e.filename2)
 
-    
+    @classmethod
+    def copy_file(cls,origin_path: str,destination_path: str) -> dict :
+            """
+            copy_file, copia los archivos de la ruta origen a una ruta destino, si el archivo ya existe en la ruta de destino este se remplazara
 
+            Args:
+                origin_path (str): Ruta de origen de los archivos
+                
+                origin_path (str): Ruta de destino de los archivos
+
+            Returns:
+                Dict: Retorna diccionario con mensaje del resultado de la operacion, tipo de resultado (successful,warning) y un estado True si la operacion finalizo con exito o False en caso contrario 
+            
+            Note:
+                -Si el archivo ya existe en la ruta de destino este se remplazara
+            """
+            try:
+                if not os.path.exists(destination_path):
+                    os.makedirs(destination_path)
+                
+                if not os.path.exists(origin_path):
+                    return  Mensage.build_message(id_mesage = 1,part1_mesage = origin_path)
+
+                files = os.listdir(origin_path) 
+                
+                for file in files:
+                    shutil.copy(os.path.join(origin_path, file),
+                                os.path.join(destination_path, file)
+                                ) 
+                    
+                return  Mensage.build_message(id_mesage = 0)      
+            except Exception as e:
+                return Mensage.build_message(0,str(e.args[1]),e.filename,e.filename2)
